+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "core-js": "^3.38.1",
     "forkify": "file:",
+    "fractional": "^1.0.0",
     "parcel": "^2.12.0",
     "regenerator-runtime": "^0.14.1"
   },
